# doc/yaml/api.yaml
openapi: 3.0.3
info:
  title: "Squealer Backend"
  description: "Backend API for all Squealer Products"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
  /users:
    get:
      tags:
        - User
      description: Get all users.
      operationId: getUsers
      parameters:
        - name: page
          in: query
          description: Paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
        - name: handle
          in: query
          description: Handle filter. Will return users whose handle contains the given string.
          schema:
            type: string
            example: userhandle99
        - name: admin
          in: query 
          description: Get only Admin accounts (true) or non admin accounts (false)
          schema:
            type: boolean
            example: True
        - name: accountType
          in: query 
          description: Filter by account type 
          schema:
            type: string
            example: pro 
            enum:
              - user 
              - pro 
        - name: handleOnly
          in: query 
          description: Return an array of handles, not of user records. 
          schema:
            type: boolean
            example: False 
            default: False 
      security:
        - adminAuth: []
      responses:
        '200':
          description: Request successfull
          content:
            application/json:
              schema:
                type: object 
                properties:
                  results: 
                    type: array
                    description: Found results.
                    items:
                      anyOf:
                        - $ref: '#/components/schemas/User'
                        - $ref: '#/components/schemas/UserPro'
                  pages: 
                    type: integer
                    minimum: 1
                    description: Number of available pages with the given search settings.
                    
        '401':
          description: Unauthorized operation
  /users/{handle}:
    put:
      tags:
        - User
      description: Create a new user with the selected handle (min 3 characters). It will return the user object and the token used for all further authorized comunication.
      operationId: createUser
      requestBody:
        description: New user
        required: True
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
      parameters:
        - name: handle
          in: path
          description: The new user's Handle
          required: True
          schema: 
            type: string
            minLength: 3
            example: userHandle123
      responses:
        '200':
          description: User Created, the JWT token is in the body.
          content:
            application/json:
              schema:
                allOf:
                  - anyOf:
                    - $ref: '#/components/schemas/User'   
                    - $ref: '#/components/schemas/UserPro'   
                  - $ref: '#/components/schemas/TokenInfo'   
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle or email
          content:
            application/json:
                schema:
                  type: object
                  properties:
                    invalid: 
                      type: array
                      description: A list of data that was invalid
                      items:
                        type: string
                        enum:
                          - handle
                          - email
                      minItems: 1
    delete:
      tags:
        - User
      description: Delete the user with the given handle
      operationId: deleteUser
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      security:
        - adminAuth: []
        - userAuth: []
      responses:
        '200':
          description: User Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
    get:
      tags:
        - User
      description: Get the user with the given handle.
      operationId: getUser
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      security:
        - adminAuth: []
        - userAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/UserPro'
        '401':
          description: Unauthorized operation.
        '409':
          description: Invalid handle
    post:
      tags:
        - User
      description: Change user object. Admins only can change the characters the blocked field.
      operationId: writeUser
      requestBody:
        description: Data to be modified
        required: True
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/WritableUserData'
                
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: User changed
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/User'
                  - $ref: '#/components/schemas/UserPro'
        '401':
          description: Unauthorized operation. To promote a user to admin one needs to provide an admin token, to set the smm one needs to be a pro user.
        '409':
          description: Invalid handle
  /users/{handle}/smm:
    post: 
      tags:
        - "User"
      description: Add/remove/change user's SMM
      operationId: changeSmm
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      requestBody:
        description: Operation Parameters 
        required: True
        content:
          application/json:
            schema:
              type: object 
              properties:
                operation: 
                  type: string
                  enum: 
                    - change 
                    - remove 
                smm: 
                  type: string 
                  description: The smm's handle, will be ignored if the operation is remove.
              required: [operation]
      responses:
        200: 
          description: Ok 
        401: 
          description: Unauthorized
        409: 
          description: Invalid Handle 
        500: 
          description: Internal server error
  /users/{handle}/managed:
    post: 
      tags:
        - "User"
      description: Remove some managed accounts.
      operationId: changeManaged
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      requestBody:
        description: An array of managed accounts to delete
        required: True
        content:
          application/json:
            schema:
              type: object 
              properties:
                users: 
                  type: array
                  minLength: 1
                  items:
                    type: string 
                    example: userHandle123
                  description: Array of users handles
              required: [users]
      responses:
        200: 
          description: User found
          content:
            application/json:
              schema:
                type: object 
                properties:
                  handle: 
                    type: string 
                    description: The user's handle 
                  charLeft:
                    $ref: '#/components/schemas/CharactersSchema'
        401: 
          description: Unauthorized
        409: 
          description: Invalid Handle 
        500: 
          description: Internal server error
    get: 
      tags:
        - "User"
      description: Get managed accounts.
      operationId: getManaged
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        200: 
          description: Ok 
        401: 
          description: Unauthorized
        409: 
          description: Invalid Handle 
        500: 
          description: Internal server error
...
  /users/{handle}/messages/:
    get:
      tags:
        - Message
        - User
      description: Get all the messages of a specific user
      operationId: getUserMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: page
          in: query
          description: Paginate results. A non-positive value disables pagination.
          schema:
            type: integer
            default: 1
            example: 3
        - name: results_per_page
          in: query
          description: Number of results to return in each query.
          schema:
            type: integer
            default: 100 
            minimum: 1
            example: 50
        - name: popular
          in: query
          description: Get only pupular messages, value determines the timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages, value determines the timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages, value determines the timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name and if it starts with
            a \# it will be interpreted as a keyword channel. Otherwise this field will be ignored.
          schema:
            type: string 
            example: ['@feltri', '§canalebello67']
        - name: answering
          in: query 
          description: Will only return messages answering to the given message's ignored.
          schema: 
            type: string 
            example: 653a84466bd663444681acfc
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
    post:
      tags:
        - Message
      description: Post a new Message. This operation will also adjust the poster's character quotas,
        if there are not enough characters it will fail with error 418.
      operationId: postUserMessage
      security:
        - adminAuth: []
        - userAuth: []
      requestBody:
        description: message to be posted
        required: True
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                  properties:
                    text: 
                      type: string
                      example: This is an example squeal that contains very interesting info.
                    image:
                      type: string
                      format: binary
                      description: An image
                dest:
                  type: array
                  items:
                    type: string
                    description: The handle for single individuals, the channel name for channels
                  example: ['@gianni', '@alberto', '§canalebello']
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Message posted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation.
        '409':
          description: Invalid handle.
        '418': 
          description: User did not have enough characters left.
          
    delete:
      tags:
        - Message
      description: Delete all messages from the user with the given handle
      operationId: deleteUserMessages
      security:
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Messages Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /users/{handle}/messages/{id}:
  
    delete:
      tags:
        - User
        - Message
      description: Delete the message with the given id from the user with the given handle.
      operationId: deleteMessage
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int128
      responses:
        '200':
          description: Message Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
    post:
      tags:
        - User
        - Message
      description: Change message data. Currently only the reactions can be changed by an admin.
      operationId: postMessage
      security:
        - adminAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int64
      requestBody:
        required: True
        description: The data to be changed
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/MessageImpressions'
                - type: object 
                  properties:
                    text: 
                      type: string 
                      description: Edited Message's text,
      responses:
        '200':
          description: Message Modified
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /messages/up/{id}:
  
    delete:
      tags:
        - User
        - Message
      description: Delete a positive reaction.
      operationId: deletePositiveReaction
      security:
        - userAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int128
      responses:
        '200':
          description: Reaction Deleted
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid id or No reaction to delete.
    post:
      tags:
        - User
        - Message
      description: Add a positive reaction to a message.
      operationId: postMessage
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          description: The message's id
          required: True
          schema:
            type: integer
            format: int64
      requestBody:
        required: True
        description: The data to be changed
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/MessageImpressions'
                - type: object 
                  properties:
                    text: 
                      type: string 
                      description: Edited Message's text,
      responses:
        '200':
          description: Message Modified
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
...
  /user/{handle}/subscription:
    post: 
      tags:
        - User 
      operationId: changeSubscription
      description: Set user subscription.
      parameters:
        - name: handle
          in: path
          description: User's handle
          required: true
          schema: 
            type: string 
            example: handle1 
        - name: proPlanName
          in: query 
          required: true 
          description: Name of the new pro plan 
          schema:
            type: string 
            example: Montly subscription plan 
        - name: autoRenew
          in: query 
          required: true
          description: Wether or not to auto renew the subscription when it expires 
          schema:
            default: true
            type: boolean 
            example: true 
      responses:
        '200':
          description: Subscription changes
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid plan name
    delete: 
      tags:
        - User 
      operationId: deleteSubscription
      description: Delete user subscription. 
      parameters:
        - name: handle
          in: path
          description: User's handle
          required: true
          schema: 
            type: string 
            example: handle1
      responses:
        '200':
          description: Subscription deleted
        '401':
          description: Unauthorized operation  
  /users/{handle}/messages/stats:
    get:
      tags:
        - Message
      description: Get users's message statistics
      operationId: getMessagesStats
      security:
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: popular
          in: query
          description: Get only pupular messages, value determines the timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages, value determines the timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages, value determines the timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name and if it starts with
            a \# it will be interpreted as a keyword channel. Otherwise this field will be ignored.
          schema:
            type: string 
            example: ['#sanremo1312', '@feltri', '§canalebello67']
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: object
                properties:
                  positive: 
                    type: number 
                    minimum: 0
                    description: Total number of likes recieved on filtered messages.
                  negative: 
                    type: number 
                    minimum: 0
                    description: Total number of dislikes recieved on filtered messages.
                  total: 
                    type: number 
                    minimum: 0
                    description: Total number of messages matched.
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /users/{handle}/joined:
    get:
      tags:
        - User
        - Channel
      description: Get channels where the user is a member
      operationId: getJoinedChannels
      security:
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Operation successful.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
  /users/{handle}/editor:
    get:
      tags:
        - User
        - Channel
      description: Get channels where the user is an editor
      operationId: getEditorChannels
      security:
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
      responses:
        '200':
          description: Operation successful.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
...
  /messages/user/{handle}: 
    get:
      tags:
        - Message
        - User
      description: Get all the messages of a specific user
      operationId: getUserMessages
      security:
        - adminAuth: []
        - userAuth: []
      parameters:
        - name: handle
          in: path
          description: The user's handle
          required: True
          schema: 
            type: string
            example: userHandle123
        - name: page
          in: query
          description: To paginate results. A non-positive value disables pagination.
          schema:
            type: integer
        - name: popular
          in: query
          description: Get only pupular messages, value determines the timeframe. If used together with unpopular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: unpopular
          in: query
          description: Get only unpupular messages, value determines the timeframe. If used together with popular 
            it will perform an or.
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: controversial
          in: query
          description: Get only controversial messages, value determines the timeframe. 
          schema:
            type: string 
            enum:
              - today
              - week 
              - month 
              - year
              - all
        - name: risk
          in: query
          description: Get only messages at risk of being controversial, popular or unpopular. 
          schema:
            type: string 
            enum:
              - controversial
              - popular 
              - unpopular
        - name: before
          in: query
          description: Get only messages dated before given timestamp.
          schema:
            type: string 
            format: date-time
        - name: after
          in: query
          description: Get only messages dated after given timestamp.
          schema:
            type: string 
            format: date-time
        - name: dest
          in: query
          description: Get only messages addressed to values. If value starts with @ it will be interpreted
            as a handle, if it starts with a § it will be interpreted as a channel name and if it starts with
            a \# it will be interpreted as a keyword channel. Otherwise this field will be ignored.
          schema:
            type: string 
            example: ['#sanremo1312', '@feltri', '§canalebello67']
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized operation
        '409':
          description: Invalid handle
...
components:
  schemas:
...
    User:
      allOf:
        - $ref: '#/components/schemas/UserHandle'
        - $ref: '#/components/schemas/UserInfo'
        - $ref: '#/components/schemas/AdminPriviledges'
        - $ref: '#/components/schemas/UserChannelsInfo'
        - type: object
          properties:
            charLeft:
              $ref: '#/components/schemas/CharactersSchema'
            meta:
              $ref: '#/components/schemas/UserMetaSchema'
            blocked:
              type: boolean
              description: A blocked user cant write messages
            lastLoggedin:
              type: string
              format: date 
              description: User last login date-time 
            subscription:
              type: object
              required: [proPlan, expires, autoRenew]
              properties:
                expires:
                  type: string
                  format: date
                  description: The expiry date
                autoRenew:
                  type: boolean
                  description: Wether to automatically renew the subscription or not.
                proPlan: 
                  $ref: '#/components/schemas/ProPlan'
          required: [meta, blocked, lastLoggedin]
    UserPro:
      allOf:
        - $ref: '#/components/schemas/User'
        - $ref: '#/components/schemas/SMMData'
        - type: object
          required: [subscription]
          properties:
            subscription:
              type: object
              required: [proPlan, expires, autoRenew]
              properties:
                expires:
                  type: string
                  format: date
                  description: The expiry date
                autoRenew:
                  type: boolean
                  description: Wether to automatically renew the subscription or not.
                proPlan: 
                  $ref: '#/components/schemas/ProPlan'
...
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires Requires a JWT token generated with user credentials. Even with a correctly formed JWT, a user can only alter their own account.
    proAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: Requires a JWT token generated with pro user credentials.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Requires a JWT token generated with admin credentials.      